fix(ios): register embedded SVG assets so icons render on iOS
Move ZedraAssets (rust-embed + AssetSource impl) from android_app.rs
to shared lib.rs and call Application::new().with_assets() in
gpui_app.rs. Without this, GPUI's SVG renderer had no asset source
and silently skipped all svg().path() calls.
